pyproject.toml
[project]
name = "bertopic-profiap"
version = "0.1.0"
description = ""
authors = [
    {name = "Thiago Duarte",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "bertopic (>=0.17.0,<0.18.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "umap-learn (>=0.5.7,<0.6.0)",
    "scikit-learn (>=1.1)",
    "spacy (>=3.8.7,<4.0.0)",
    "octis (>=1.14.0,<2.0.0)",
    "xlrd (>=2.0.1,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "datamapplot (>=0.6.3,<0.7.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "nbclient (>=0.10.2,<0.11.0)",
    "itables (>=2.4.3,<3.0.0)",
    "openai (>=1.93.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "bertopic_profiap", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
